Mimic fixes
The mimic now properly scales it's BTC reward on the floors where enemies get stronger. They also now display properly in the viewport, instead of being invisible (dead giveaway that a chest was a mimic). Mimic chests can also be destroyed with C-4 now, and give you the same BTC drop that you are rewarded with when defeating them in a regular battle.