Changes to LUK effects + Gambler fix
Increasing LUK now increases your chances at landing a 12 in the gambler diceroll minigame. I also noticed in the recent past that landing a 12 (which was also COMICALLY rare btw) would not trigger a win state. Something must have broke in an update sometime down the line, but that's been fixed. Also nerfed the crit chance rate when leveling LUK from 2% down to 1%.