Party size + attempt limit, new ring, PRS stat scaling changes, various fixes
You can now have a maximum of 3 members in your party, and by default, can only make two attempts at persuasion per battle. I've added a ring (Ring of Amplified Audio) that increases the attempt limit to 3. This won't spawn in chests and can only be purchased from the Merchant.

I've also adjusted the PRS stat scaling so that it caps out at 100. Meaning, for example, if you had 12 PRS, your chances at persuading an enemy would be 12%, and so on. I've increased the base PRS stat to 15 to reflect this change.

I've also fixed a couple of those bugs that were brought up. Player input is now locked while gambling and the "You died on floor X" message has been moved to the battle log.

This will close #56 , #55 , and #51 .