Integrated title screen, new logo + art, etc
This was a nightmare to get working the way I wanted it to, but we now have a title screen engrained into the main game file itself. I've also created created a new logo, new artwork, and there's a fancy screen wipe effect when you enter the game.
Additionally, I've added a quickstart toggle setting in the options menu that allows you to automatically skip the title screen upon booting the game up.

Also renamed game.htm to game.html just because that is more convenient for me.