LUK scaling changes
Gambling chances used to scale in a strange way, resulting in the LUK stat softcapping at a rather low value, allowing you to easily win the diceroll relatively early on.

This now behaves like the PRS stat where the chances are on a 0-100% scale in accordance to your stat number. For example, if your LUK stat is at 20, your chances at winning the diceroll minigame are 20%. The BTC gain from chests has been left untouched here.