Current music + Gamepad connection status bar
There are now two thin boxes at the bottom of the interface displaying the current song being played, as well as a status indicator for whether or not a gamepad is connected.

Currently I am having issues actually aligning these boxes with the rest of the interface, help will be needed here, as I am seemingly lacking understanding of this whole interface grid system that's been semi-recently devised.