Content centering
Included here is an index.html that will display the game in its full glory centered in the middle of the window. It will also dynamically zoom the content and fill the window's height. This will especially look nice if I/we ever decided to add artwork borders here. I've also removed all padding from the area surrounding the content in game.html as well.